fix(i): Move cache action after goreleaser action (#2395)
## Relevant issue(s)

Resolves #2394 

## Description

This PR moves the cache action in the release workflow to be after the
goreleaser action. After the last release, we changed the cache action
to use `action/cache/save` instead of just `action/cache`. The
difference is that `action/cache` acts like a defer function and gets
actioned at the end of a job. `action/cache/save` is meant to be more
granular and tries to save when called. This caused a situation where it
tried to save in the cache a path that didn't exist.