fix(i): Correctly handle inverse lenses from mult. sources (#2530)
## Relevant issue(s)

Resolves #2433

## Description

Correctly handles inverse lenses from multiple sources.

PR includes a new test for verifying that inversing across multiple
versions does work, it is not relevant to the bug fix, but I wrote it to
make sure I want reading the code wrong (as I first did when creating
the issue).

The bug is not testable atm, as this would only have become a problem
when introducing multi-source collections (atm both `next` and
`previous` slices are always empty before calling this line). Was likely
copy-paste code.