refactor: DAG sync and move merge outside of net package (#2658)
## Relevant issue(s)

Resolves #2624 

## Description

This PR simplifies the DAG sync process withing the `net` package and
moves the merge functionality to the `db` package. The merge is now
initiated via an event channel.

Note: I did a search and replace for `SchemaVersionId` to
`SchemaVersionID`. It's in its own commit. I've also remove the
`tests/integration/net/order` tests as they are now annoying to maintain
an will become even more irrelevant when we refactor the WaitForSync
functionality of our test framework.

Another note: I've reduced the severity of the race condition on my Mac.
We had a lot of leaking go routines and what is left of them is
WaitForSync methods that sometimes seem to leak and also badger cache
and libp2p transport that seem to leak go routines on close but I'm not
sure how to handle these last two.