refactor: Change local_acp implementation to use acp_core (#2691)
## Relevant issue(s)


Also the related issue won't close upon merge, nor is it picked up by
github at the moment, because you just have #2694 in the PR description.

So please don't forget to add the resolving key word (close, closes,
closed, fix, fixes, fixed, resolve, resolves, resolved) with it.

So you can put for example:

Resolve #2694

## Description

This PR changes the underlying implementation of the `ACPLocal`
Reference Monitor from SourceHub to ACP Core.
The changes aim to allow the WASM build which was broken since the
inclusion of SourceHub as a dependency.

Notable changes from this refactor include:
- Replace SourceHub as a dependency for acp_core for ACP Local 
- The field `name` in a `Policy` is now correctly marked as required.
This was a consequence of an issue on SourceHub side which allowed
unnamed policies to be included. Therefore, in order to correct the
tests, a name was given to all policies in the codebase.
- The previous implementation of ACP Local made use of a special notion
of account which had no sequence numbers. As such, re-submitting a
Policy generated the same ID. This behavior isn't compatible with how
SourceHub deals with IDs and the new implementation using ACP Core
reflects that. That is, every created Policy, independent of its
payload, will have unique IDs.
- Updated the `Identity` struct to generate a DID instead of a SourceHub
address

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Test suite ie `make test`

Specify the platform(s) on which this was tested:
- Arch Linux