feat: Ability to generate a new identity (#2760)
## Relevant issue(s)
Resolves #2554 

## Description
Now that the `core_acp` work is merged the identity generation becomes
much more simple, with a `did` instead of `source address`. This util
command can be handy to quickly generate new identity pairs with their
did.

### Demo

Issue the command:
```
$ defradb identity new
```

Output:
```json
{
  "PrivateKey": "8wsOdJZCg06ISJxBqUyrD1ovDF5WgJJrSup2i4avnuU=",
  "PublicKey": "BG6dw3UW3rv+IiTXfL0LXvACw7Qz3DSBy1KmS/eSNrsrRkToBsize7zp+xwFrelPylOPgZguG7ZmirAtTaY89ow=",
  "DID": "did:key:z7r8oqgz72xy7Z7V5WGF21iWc7EY8vCpjZA1nGugoKjGeppeMkqeD7FV5AxEQAnciBz5vk5c2zrTZytofJNM8ZxyBzSpT"
}

```


## How has this been tested?
- cli test
- unit tests

Specify the platform(s) on which this was tested:
- WSL2