fix(i): Tests with primary doc patch update events (#2850)
## Relevant issue(s)

Resolves #2849

## Description

This PR fixes an issue with primary document patch updates. We also now
assert that the document updates match the expected updates. We can most
likely remove the `events` test package now that every test asserts the
expected events.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

`make test`

Specify the platform(s) on which this was tested:
- MacOS