feat: Add materialized views (#3000)
## Relevant issue(s)

Resolves #2951

## Description

Adds materialized views. Also makes materialized views the default (see
discord discussion).

The caching behaviour of views in tests is now selected via an
environment variable, meaning (with the exception of a few specific
examples) a test with a view will test both cacheless and materialized
variants - in the CI this adds a new dimension to the matrix, although
materialized views are only executed using the simple settings (in-mem
store, go client, etc) for now.