chore: Make keyring non-interactive (#3026)
## Relevant issue(s)

Resolves #2995 

## Description

This PR makes the keyring interactions non-interactive by requiring the
keyring password to be set as an environment variable secret. It also
adds support for that secret to be stored in a `.env` file in the
working directory or in a file at a path defined by the `--secret-file`
flag.

Making the keyring non-interactive is necessary to support automated
deployments.