test: Allow soft-referencing of Cids in tests (#3176)
## Relevant issue(s)

Resolves #3172

## Description

Allows soft-referencing of Cids in tests.

When encountering an unpleasant change-detector problem in
https://github.com/sourcenetwork/defradb/pull/3173 I realised we had no
need to reference Cids based on index, and we really only ever cared
that they were valid, and either not the same as another cid, or the
same.

This change allows us to soft-reference them without ever really caring
about where they came from. This allows us to give them descriptive ids,
and avoids having to have the test framework from having to fetch them
independently from the DB - something that raises questions over whether
we are testing the test framework or the production code, especially in
the change detector.

The system can be extended to other types fairly easily. 

## How has this been tested?

I had a fairly long play around giving it bad ids in tests to make sure
it would actually fail when appropriate.