fix: Resolve CORS errors in OpenAPI tab of Playground (#3263)
## Relevant issue(s)

Resolves #2458

## Description

There was a bug in the Playground's OpenAPI tool. This tool would list
the URL of the endpoints as localhost instead of 127.0.0.1 (which is
what the startup Defradb message shows.) The result was that attempting
to execute one of the queries returned a CORS error.

This was a one-liner fix. I adjusted openapi.go to use
`http://127.0.0.1:9181/api/v0` as a string literal for the URL, instead
of `http://localhost:9181/api/v0`

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in nts,

## How has this been tested?

The platform(s) on which this was tested:
- Windows