docs(i): Replace broken DQL link in README (#3274)
## Relevant issue(s)

Resolves #3269 

## Description

This is a change to update a broken link in the README.
This changes the target URI to point to an extant page in the docs.
FROM:

https://docs.source.network/references/query-specification/query-language-overview
TO:

https://docs.source.network/defradb/references/query-specification/query-language-overview

NO CODE CHANGES

## Tasks

- [NA] I made sure the code is well commented, particularly
hard-to-understand areas.
- [X] I made sure the repository-held documentation is changed
accordingly.
- [X] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [NA] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

I manually followed the old link and the new link. Old was a broken
link. New is a proposed correct link.

Specify the platform(s) on which this was tested:
- MacOS