fix: Add support for operationName and variables in HTTP GET (#3292)
## Relevant issue(s)

Resolves #3153

## Description

I have made it so that operationName and variables are supported by HTTP
GET requests. I did this by modifying the `ExecRequest` function inside
`http/handler_store.go` such that `operationName` and `variables`
parameters are extracted if they are present.

I added two new tests to `http/handler_store_test.go` which test this
functionality. See:

`TestExecRequest_WithValidQuery_HttpGet_WithOperationName_OmitsErrors`
and
`TestExecRequest_HttpGet_WithVariables_OmitsErrors`

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).

## How has this been tested?

The platform(s) on which this was tested:
- Windows