fix: Adjust OpenAPI index POST example request body (#3268)
## Relevant issue(s)

Resolves #2459

## Description

On the OpenAPI tool of the playground, the Index POST tab contained an
example body that was incorrect. It listed ID as a field which could be
set, which was untrue (this value is automatically generated, and trying
to pass one in throws an error.) To adjust this, I have created a new
description structure called `IndexCreateRequestDescription` which is
meant to be used for this type of HTTP request. To go alongside this
change, a new schema type, `indexCreateRequestSchema` was also created.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).


## How has this been tested?

The platform(s) on which this was tested:
- Windows