chore: Import defradb/errors for Join (#3343)
## Relevant issue(s)

Resolves #3334 

## Description

In many locations in the code base, we were importing the `"errors"`
package in order to utilize the `Join` function. However, we now have
this functionality as part of our own errors package. So throughout the
codebase import statements, `"errors"` has been changed to
`"github.com/sourcenetwork/defradb/errors"`

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [ ] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- Windows