feat: Add contraints and embedding directives (#3405)
## Relevant issue(s)

Resolves #3350
Resolves #3351

## Description

Sorry for having 3 technically separate features as part of 1 PR. The
reason for this is that I started working on embeddings and the size
contraint was initially part of the embedding. Since we discussed that
it could be applied to all array fields (and later even to String and
Blob), I extracted it into a contraints directive that has a size
parameter (more contraints can be added in the future). Furthermore,
embeddings returned from ML models are arrays of float32. This caused
some precision issues because we only supported float64. When saving the
float32 array, querying it would return an float64 array with slight
precision issues. I decided to add the float32 type.

You can review the first commit for contraint and embedding related code
and the second commit for the float related changes. Some float stuff
might have leaked in the first commit. Sorry for this. I tried hard to
separate the float32 related changes.

Note that the `gql.Float` type is now `Float64` internally.
```graphql
type User {
  points: Float
}
```
is the same as
```graphql
type User {
  points: Float64
}
```

The embedding generation relies on a 3rd party package called
`chromem-go` to call the model provider API. As long as one of the
supported provider API is configured and accessible, the embeddings will
be generated when adding new documents. I've added a step in the test
workflow that will run the embedding specific tests on linux only (this
is because installation on windows and mac is less straight forward)
using Ollama (because it runs locally).

The call to the API has to be done synchronously otherwise the docID/CID
won't be representative of the contents. The only alternative would be
for the system to automatically update the document when returning from
the API call but I see that as a inferior option as it hides the update
step from the user. It could also make doc anchoring more complicated as
the user would have to remember to wait on the doc update before
anchoring the doc at the right CID.

We could avoid having embedding generation support and let the users do
that call themselves and store the embedding vectors directly. However,
having it as a feature allows us to support RAG vector search which
would let users get started with AI with very little work. This seems to
be something our partners are looking forward to.

I don't see the 3rd party API call inline with a mutation as a problem
since this is something that has to be configured by users and users
will expect the mutation calls to take more time as a result.

If you're interested in running it locally, install Ollama and define a
schema like so
```graphql
type User {
    name: String
    about: String
    name_v: [Float32!] @constraints(size: 768) @embedding(fields: ["name", "about"], provider: "ollama", model: "nomic-embed-text",  url: "http://localhost:11434/api") // contraint is optional and localhost:11434 is the default port for Ollama
}
```

Next steps: 
 - Support templates for the content sent to the model.
- Add the `_similarity` operation to calculate the cosine similarity
between two arrays.