refactor: Make validator return joined errors (#3513)
## Relevant issue(s)

Resolves #2568

## Description

Validator functions currently were returning singular errors when
encountered. With this change, they will instead return a list of errors
that are encountered, by using the Join() function on the errors.

To accommodate this change, additional logic was added to some validator
functions to check for nil values being passed in. This is to avoid and
prevent nil pointer exceptions.

This is still in progress.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- Windows