test: Add example CLI integration tests (#3550)
## Relevant issue(s)

Resolves #3569

## Description

Adds the example CLI integration tests using new test framework.

Largely as seen in the demo last week, but with a little more polish. I
think it is good to commit these now and then we can expand them as/when
we feel like, and use it as an example for other user-facing stuff (like
http).

I have not added the txn-commit multiplier to the CI, as the workflow
test files are fairly coupled to executing all tests in the entire
repository and I wanted to limit scope here.

## Tasks

- [x] The testing repo dependency is currently a local, relative, path
and needs to be pointed at https://github.com/sourcenetwork/testing when
https://github.com/sourcenetwork/testing/pull/1 gets merged.
- [x] Link this to a proper issue if/when
https://github.com/sourcenetwork/testing/pull/1 is merged