feat: Make replicator retry intervals configurable (#3574)
## Relevant issue(s)

Resolves #3073

## Description

This adds some functionality for configuring the durations of time
between the replicator retries. Originally, a hardcoded set of values
was defined in `internal/db/config.go`. Now, this is a config option
inside `cli/config.go` called `replicator.retryIntervals`.

Additionally, there is a config flag, `replicator-retry-intervals` that
can be passed in as a slice of integers via CLI:

`defradb start --replicator-retry-intervals=10,20,66`

Further, there are some new unit tests included to test for ways in
which the flag might throw errors on bad input.

## Tasks

- [x] I made sure the code is well commented, particularly
hard-to-understand areas.
- [x] I made sure the repository-held documentation is changed
accordingly.
- [x] I made sure the pull request title adheres to the conventional
commit style (the subset used in the project can be found in
[tools/configs/chglog/config.yml](tools/configs/chglog/config.yml)).
- [x] I made sure to discuss its limitations such as threats to
validity, vulnerability to mistake and misuse, robustness to
invalidation of assumptions, resource requirements, ...

## How has this been tested?

Specify the platform(s) on which this was tested:
- Windows